there no changes here

+++ DashboardFitness/src/main/java/ma/emsi/dashboardfitness/repositories/IEntrainementRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 
 public interface IEntrainementRepository extends JpaRepository<Entrainement, Long> {
     Entrainement findByNom(String name);
-    Entrainement findByDuree(int duree);
+    Entrainement findByDuree(long duree);
     Entrainement findByNutrition (Nutrition nutrition);
     @Query("SELECT e FROM Entrainement e WHERE e.imcMin <= :imc AND e.imcMax >= :imc")
     Entrainement findByIMCRange(double imc);
